Add a feature to round building floats
The new flag will round locations and rotations of buildings to 1/64th
of a tile and 1 degree respectively. It decreases blueprint size by 30%
or so. Also, the game tends to put "on-grid" buildings an epsilon
off-grid, so this helps perfectly align them.

We explicitly give no guarantees for the changed blueprints to work.

Kudos to @Neutron3529 for suggesting and initially implementing this
feature.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>